move inventory package out of entity package
