+++ pizzaserver-server-api/src/main/java/io/github/pizzaserver/api/blockentity/impl/BlockEntityContainer.java
@@ -3,7 +3,7 @@ package io.github.pizzaserver.api.blockentity.impl;
 import com.nukkitx.protocol.bedrock.data.inventory.ContainerType;
 import io.github.pizzaserver.api.Server;
 import io.github.pizzaserver.api.blockentity.BaseBlockEntity;
-import io.github.pizzaserver.api.entity.inventory.BlockEntityInventory;
+import io.github.pizzaserver.api.inventory.BlockEntityInventory;
 import io.github.pizzaserver.api.event.type.inventory.InventoryOpenEvent;
 import io.github.pizzaserver.api.player.Player;
 import io.github.pizzaserver.api.utils.BlockLocation;
