+++ pizzaserver-server-api/src/main/java/io/github/pizzaserver/api/event/type/inventory/InventoryCreativeDestructionEvent.java
@@ -1,7 +1,7 @@
 package io.github.pizzaserver.api.event.type.inventory;
 
 import com.nukkitx.protocol.bedrock.data.inventory.ContainerSlotType;
-import io.github.pizzaserver.api.entity.inventory.Inventory;
+import io.github.pizzaserver.api.inventory.Inventory;
 import io.github.pizzaserver.api.item.Item;
 import io.github.pizzaserver.api.player.Player;
