+++ pizzaserver-server-api/src/main/java/io/github/pizzaserver/api/event/type/inventory/InventoryOpenEvent.java
@@ -1,6 +1,6 @@
 package io.github.pizzaserver.api.event.type.inventory;
 
-import io.github.pizzaserver.api.entity.inventory.Inventory;
+import io.github.pizzaserver.api.inventory.Inventory;
 import io.github.pizzaserver.api.player.Player;
 
 /**
