+++ pizzaserver-server-api/src/main/java/io/github/pizzaserver/api/inventory/BlockEntityInventory.java
@@ -1,4 +1,4 @@
-package io.github.pizzaserver.api.entity.inventory;
+package io.github.pizzaserver.api.inventory;
 
 import io.github.pizzaserver.api.blockentity.BlockEntity;
