+++ pizzaserver-server-api/src/main/java/io/github/pizzaserver/api/inventory/PlayerInventory.java
@@ -1,4 +1,4 @@
-package io.github.pizzaserver.api.entity.inventory;
+package io.github.pizzaserver.api.inventory;
 
 import io.github.pizzaserver.api.item.Item;
 import io.github.pizzaserver.api.player.Player;
