+++ pizzaserver-server/src/main/java/io/github/pizzaserver/server/ImplServer.java
@@ -9,8 +9,8 @@ import io.github.pizzaserver.api.blockentity.BlockEntityRegistry;
 import io.github.pizzaserver.api.entity.Entity;
 import io.github.pizzaserver.api.entity.EntityRegistry;
 import io.github.pizzaserver.api.entity.boss.BossBar;
-import io.github.pizzaserver.api.entity.inventory.BlockEntityInventory;
-import io.github.pizzaserver.api.entity.inventory.EntityInventory;
+import io.github.pizzaserver.api.inventory.BlockEntityInventory;
+import io.github.pizzaserver.api.inventory.EntityInventory;
 import io.github.pizzaserver.api.event.EventManager;
 import io.github.pizzaserver.api.item.CreativeRegistry;
 import io.github.pizzaserver.api.item.Item;
@@ -27,9 +27,9 @@ import io.github.pizzaserver.server.block.ImplBlockRegistry;
 import io.github.pizzaserver.server.blockentity.ImplBlockEntityRegistry;
 import io.github.pizzaserver.server.entity.ImplEntityRegistry;
 import io.github.pizzaserver.server.entity.boss.ImplBossBar;
-import io.github.pizzaserver.server.entity.inventory.ImplBlockEntityInventory;
-import io.github.pizzaserver.server.entity.inventory.ImplEntityInventory;
-import io.github.pizzaserver.server.entity.inventory.InventoryUtils;
+import io.github.pizzaserver.server.inventory.ImplBlockEntityInventory;
+import io.github.pizzaserver.server.inventory.ImplEntityInventory;
+import io.github.pizzaserver.server.inventory.InventoryUtils;
 import io.github.pizzaserver.server.event.ImplEventManager;
 import io.github.pizzaserver.server.item.ImplCreativeRegistry;
 import io.github.pizzaserver.server.item.ImplItemRegistry;
