+++ pizzaserver-server/src/main/java/io/github/pizzaserver/server/inventory/InventoryUtils.java
@@ -1,4 +1,4 @@
-package io.github.pizzaserver.server.entity.inventory;
+package io.github.pizzaserver.server.inventory;
 
 import com.nukkitx.protocol.bedrock.data.inventory.ContainerSlotType;
 import com.nukkitx.protocol.bedrock.data.inventory.ContainerType;
