+++ pizzaserver-server/src/main/java/io/github/pizzaserver/server/player/ImplPlayer.java
@@ -11,7 +11,6 @@ import com.nukkitx.protocol.bedrock.data.entity.EntityFlags;
 import com.nukkitx.protocol.bedrock.data.inventory.ItemData;
 import com.nukkitx.protocol.bedrock.packet.*;
 import io.github.pizzaserver.api.block.Block;
-import io.github.pizzaserver.api.block.BlockID;
 import io.github.pizzaserver.api.blockentity.BlockEntity;
 import io.github.pizzaserver.api.entity.Entity;
 import io.github.pizzaserver.api.entity.boss.BossBar;
@@ -19,17 +18,12 @@ import io.github.pizzaserver.api.entity.data.attributes.Attribute;
 import io.github.pizzaserver.api.entity.data.attributes.AttributeType;
 import io.github.pizzaserver.api.entity.definition.impl.EntityCowDefinition;
 import io.github.pizzaserver.api.entity.definition.impl.EntityHumanDefinition;
-import io.github.pizzaserver.api.entity.inventory.Inventory;
-import io.github.pizzaserver.api.event.type.block.BlockStopBreakEvent;
+import io.github.pizzaserver.api.inventory.Inventory;
 import io.github.pizzaserver.api.event.type.entity.EntityDamageEvent;
 import io.github.pizzaserver.api.event.type.player.PlayerLoginEvent;
 import io.github.pizzaserver.api.event.type.player.PlayerRespawnEvent;
 import io.github.pizzaserver.api.item.CreativeRegistry;
 import io.github.pizzaserver.api.item.data.ItemID;
-import io.github.pizzaserver.api.item.impl.ItemAcaciaBoat;
-import io.github.pizzaserver.api.item.impl.ItemArrow;
-import io.github.pizzaserver.api.item.impl.ItemBlock;
-import io.github.pizzaserver.api.item.impl.ItemDiamond;
 import io.github.pizzaserver.api.level.data.Difficulty;
 import io.github.pizzaserver.api.level.world.World;
 import io.github.pizzaserver.api.level.world.data.Dimension;
@@ -52,8 +46,8 @@ import io.github.pizzaserver.server.ImplServer;
 import io.github.pizzaserver.server.entity.ImplEntity;
 import io.github.pizzaserver.server.entity.ImplEntityHuman;
 import io.github.pizzaserver.server.entity.boss.ImplBossBar;
-import io.github.pizzaserver.server.entity.inventory.BaseInventory;
-import io.github.pizzaserver.server.entity.inventory.ImplPlayerInventory;
+import io.github.pizzaserver.server.inventory.BaseInventory;
+import io.github.pizzaserver.server.inventory.ImplPlayerInventory;
 import io.github.pizzaserver.server.item.ItemUtils;
 import io.github.pizzaserver.server.level.world.ImplWorld;
 import io.github.pizzaserver.server.network.data.LoginData;
