+++ pizzaserver-server/src/main/java/io/github/pizzaserver/server/player/handlers/inventory/InventoryActionDestroyHandler.java
@@ -1,9 +1,6 @@
 package io.github.pizzaserver.server.player.handlers.inventory;
 
-import io.github.pizzaserver.api.entity.EntityItem;
-import io.github.pizzaserver.api.entity.EntityRegistry;
 import io.github.pizzaserver.api.event.type.inventory.InventoryCreativeDestructionEvent;
-import io.github.pizzaserver.api.event.type.inventory.InventoryDropItemEvent;
 import io.github.pizzaserver.api.item.Item;
 import io.github.pizzaserver.api.player.Player;
 import io.github.pizzaserver.server.network.data.inventory.InventorySlotContainer;
@@ -16,7 +13,7 @@ public class InventoryActionDestroyHandler extends InventoryActionHandler<Destro
     @Override
     protected boolean isValid(Player player, DestroyStackRequestActionDataWrapper action) {
         InventorySlotContainer sourceSlot = action.getSource();
-        return sourceSlot.exists() && action.getCountRequested() > 0;
+        return sourceSlot.exists() && action.getCountRequested() > 0 && player.isCreativeMode();
     }
 
     @Override
